construction controller

+++ Geolaxia/Model/Facility.cs
@@ -3,7 +3,7 @@
     public abstract class Facility
     {
         public virtual int Id { get; set; }
-        public virtual double ConstructionTime { get; set; }
+        public virtual int ConstructionTime { get; set; }
         public virtual Cost Cost { get; set; }
         public virtual int Level { get; set; }
         public virtual Planet Planet { get; set; }
